Re added all the debugged codes and updated gitignore. All the code seems to have been lost during the merging.

+++ Calendified/Library/logic.cpp
@@ -78,9 +78,13 @@ std::string logic::readCommand(std::string commandLine, int flipCount){
 		currentDisplayContent = displayTask.getDisplayContent();
 		indexToActOnDisplay = currentCommandReference.getIndexToBeActOn();
 		indexToActOnStorage = displayTask.getStorageIndex(currentDisplayContent,indexToActOnDisplay);
-		deleteResults = deleteItem.executeDelete(indexToActOnStorage);		
-		deleteItem.undoDelete(&undoTask);		
-		return deleteResults;
+		if(indexToActOnStorage < 0){
+			return MESSAGE_ERROR_INVALID_INDEX;
+		} else { 
+			deleteResults = deleteItem.executeDelete(indexToActOnStorage);		
+			deleteItem.undoDelete(&undoTask);		
+			return deleteResults;
+		}
 	case SEARCH:
 		displayTask.updateStorageSource();
 		displayTask.setDisplayContent(emptyTaskList);
@@ -122,10 +126,14 @@ std::string logic::readCommand(std::string commandLine, int flipCount){
 		currentDisplayContent = displayTask.getDisplayContent();
 		indexToActOnDisplay = currentCommandReference.getIndexToBeActOn();
 		indexToActOnStorage = displayTask.getStorageIndex(currentDisplayContent, indexToActOnDisplay);
-		editItem.setEditingRef(newTask);
-		editResults = editItem.executeEdit(indexToActOnStorage);
-		editItem.undoEdit(&undoTask);
-		return editResults;
+		if(indexToActOnStorage < 0){
+			return MESSAGE_ERROR_INVALID_INDEX;
+		} else {
+			editItem.setEditingRef(newTask);
+			editResults = editItem.executeEdit(indexToActOnStorage);
+			editItem.undoEdit(&undoTask);
+			return editResults;
+		}
 	case DONE:
 		displayTask.updateStorageSource();
 		currentDisplayContent = displayTask.getDisplayContent();
